MLBEDSW-10106: Fix bitflags casting issue.
An obscure cast issue occured during another implementation
where attempting to cast the bitflags object causes it to
take the implicit boolean value, rather than the explicit
unsigned cast.

 - This commit prevents such casts by removing the
   implicit boolean cast.
 - Added specific flag-test operator, resulting in
   an actual boolean value.
 - Allowed double-negation semantics (!!) to check
   for non-zero flags.

+++ ethosu/regor/architecture/architecture.hpp
@@ -1,5 +1,5 @@
 //
-// SPDX-FileCopyrightText: Copyright 2021-2024 Arm Limited and/or its affiliates <open-source-office@arm.com>
+// SPDX-FileCopyrightText: Copyright 2021-2025 Arm Limited and/or its affiliates <open-source-office@arm.com>
 //
 // SPDX-License-Identifier: Apache-2.0
 //
@@ -119,7 +119,7 @@ struct MemArea
 
     struct hash
     {
-        size_t operator()(const MemArea &memArea) const { return size_t(memArea.memory) | size_t(memArea.usage); }
+        size_t operator()(const MemArea &memArea) const { return size_t(memArea.memory) | unsigned(memArea.usage); }
     };
 };
